refactored stuff into mtvCheckoutSteps.js
